SCA-50: added likes and comments counters on the posts list. Changed counters type from int to long.

+++ app/src/main/java/com/rozdoum/socialcomponents/model/Post.java
@@ -19,8 +19,8 @@ public class Post implements Serializable{
     private long createdDate;
     private String imagePath;
     private String authorId;
-    private int commentsCount;
-    private int likesCount;
+    private long commentsCount;
+    private long likesCount;
 
     public Post() {
         this.createdDate = new Date().getTime();
@@ -74,19 +74,19 @@ public class Post implements Serializable{
         this.createdDate = createdDate;
     }
 
-    public int getCommentsCount() {
+    public long getCommentsCount() {
         return commentsCount;
     }
 
-    public void setCommentsCount(int commentsCount) {
+    public void setCommentsCount(long commentsCount) {
         this.commentsCount = commentsCount;
     }
 
-    public int getLikesCount() {
+    public long getLikesCount() {
         return likesCount;
     }
 
-    public void setLikesCount(int likesCount) {
+    public void setLikesCount(long likesCount) {
         this.likesCount = likesCount;
     }
